nagios_dont_blame True now allows arguments
So setting this ansible variable:

<pre>
nagios_dont_blame: True
</pre>

now sets this nagios.cfg setting:

<pre>
dont_blame_nrpe=1
</pre>

Which allows sending arguments to the checks

This solves #2

We also change the default to False - so that the default behavior of
this role is the same.

If you(like me) have set nagios_dont_blame: False
you will have to change it to True to keep allowing command arguments.